[4633] Move hpitt import code into a service to match rest of codebase
